Implement external editor but it doesn't draw on exit
