+++ src/app.rs
@@ -330,6 +330,9 @@ impl<'repo> App<'repo> {
             if let Some(pager) = &self.blob_pager {
                 self.external_editor = Some(ExternalEditor::new(&pager.blob, &pager.name, "emacsclient"));
                 self.mode_history.push(AppMode::ExternalEditor);
+                if let Some(external_editor) = &mut self.external_editor {
+                    external_editor.display();
+                }
             }
         }
     }
