+++ src/app/blob_pager.rs
@@ -1,6 +1,3 @@
-use std::io::Write as _;
-use std::process::Command;
-
 use git2::{Blob, Object, Repository};
 
 use ratatui::{
@@ -9,8 +6,6 @@ use ratatui::{
     widgets::{Block, Paragraph},
     Frame,
 };
-
-use tempfile::Builder;
 
 use color_eyre::Result;
 
@@ -23,7 +18,6 @@ pub struct BlobPager<'repo> {
     pub blob: Blob<'repo>,
     pub name: String,
     lines: Vec<String>,
-    content: Vec<u8>,
 }
 
 impl<'repo> BlobPager<'repo> {
@@ -40,7 +34,6 @@ impl<'repo> BlobPager<'repo> {
             blob: blob.clone(),
             name,
             lines,
-            content: content.to_owned(),
         }
     }
 
@@ -155,26 +148,6 @@ impl<'repo> Navigable<'repo> for BlobPager<'repo> {
     }
 
     fn select(&self) -> Option<(Object<'repo>, String)> {
-        match Builder::new().suffix(&self.name).tempfile() {
-            Ok(mut tempfile) => {
-                let file = tempfile.as_file_mut();
-                file.write_all(&self.content).expect("failed to write file");
-
-                let mut emacsclient = Command::new("emacsclient")
-                    .arg(tempfile.path())
-                    .spawn()
-                    .expect("failed to run emacsclient");
-
-                let ecode = emacsclient
-                    .wait()
-                    .expect("failed waiting for emacsclient to exit");
-
-                assert!(ecode.success());
-            }
-            // We should use a handlable error
-            Err(_) => panic!("unable to create temporary file"),
-        }
-
         None
     }
