+++ src/app/external_editor.rs
@@ -1,4 +1,11 @@
-use git2::{Blob, Object, Repository};
+use std::io::Write as _;
+use std::process::Command;
+
+use git2::Blob;
+
+use tempfile::Builder;
+
+use crate::tui;
 
 pub struct ExternalEditor {
     editor: String,
@@ -13,6 +20,31 @@ impl<'repo> ExternalEditor {
             editor: editor.to_string(),
             name: name.to_string(),
             content: blob.content().to_owned(),
+        }
+    }
+
+    pub fn display(&self) {
+        match Builder::new().suffix(&self.name).tempfile() {
+            Ok(mut tempfile) => {
+                tui::restore().expect("failed to restore terminal");
+                eprintln!("Opening {} with {} ...", self.name, self.editor);
+                let file = tempfile.as_file_mut();
+                file.write_all(&self.content).expect("failed to write file");
+
+                let mut emacsclient = Command::new(&self.editor)
+                    .arg(tempfile.path())
+                    .spawn()
+                    .expect("failed to run emacsclient");
+
+                let ecode = emacsclient
+                    .wait()
+                    .expect("failed waiting for emacsclient to exit");
+
+                assert!(ecode.success());
+                tui::init().expect("failed to reinit terminal");
+            }
+            // We should use a handlable error
+            Err(_) => panic!("unable to create temporary file"),
         }
     }
 }
